+++ include/GafferArnold/Private/IECoreArnold/ProceduralAlgo.h
@@ -32,62 +32,23 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#include "GafferArnold/Private/IECoreArnoldPreview/ProceduralAlgo.h"
+#ifndef IECOREARNOLD_PROCEDURALALGO_H
+#define IECOREARNOLD_PROCEDURALALGO_H
 
-#include "IECoreArnold/NodeAlgo.h"
-#include "IECoreArnold/ParameterAlgo.h"
+#include "IECoreScene/ExternalProcedural.h"
 
-#include "IECore/SimpleTypedData.h"
-#include "IECore/Version.h"
+#include "ai.h"
 
-using namespace std;
-using namespace Imath;
-using namespace IECore;
-using namespace IECoreScene;
-using namespace IECoreArnold;
-using namespace IECoreArnoldPreview;
-
-//////////////////////////////////////////////////////////////////////////
-// Internal utilities
-//////////////////////////////////////////////////////////////////////////
-
-namespace
-{
-
-#if CORTEX_COMPATIBILITY_VERSION < 10003
-
-NodeAlgo::ConverterDescription<ExternalProcedural> g_description(
-	[] ( const IECoreScene::ExternalProcedural *procedural, const std::string &name, const AtNode *parent ) {
-		return ProceduralAlgo::convert( procedural, nullptr, name, parent );
-	}
-);
-
-#else
-
-NodeAlgo::ConverterDescription<ExternalProcedural> g_description( ProceduralAlgo::convert );
-
-#endif
-
-} // namespace
-
-//////////////////////////////////////////////////////////////////////////
-// Implementation of public API
-//////////////////////////////////////////////////////////////////////////
-
-namespace IECoreArnoldPreview
+namespace IECoreArnold
 {
 
 namespace ProceduralAlgo
 {
 
-AtNode *convert( const IECoreScene::ExternalProcedural *procedural, AtUniverse *universe, const std::string &nodeName, const AtNode *parentNode  )
-{
-	AtNode *node = AiNode( universe, AtString( procedural->getFileName().c_str() ), AtString( nodeName.c_str() ), parentNode );
-	ParameterAlgo::setParameters( node, procedural->parameters()->readable() );
-
-	return node;
-}
+AtNode *convert( const IECoreScene::ExternalProcedural *procedural, AtUniverse *universe, const std::string &nodeName, const AtNode *parentNode );
 
 } // namespace ProceduralAlgo
 
-} // namespace IECoreArnoldPreview
+} // namespace IECoreArnold
+
+#endif // IECOREARNOLD_PROCEDURALALGO_H
